日志分析任务添加开关

+++ src/main/java/com/bkcc/logans/controller/TaskController.java
@@ -1,10 +1,14 @@
 package com.bkcc.logans.controller;
 
+import com.alibaba.fastjson.JSONObject;
 import com.bkcc.core.data.ViewData;
 import com.bkcc.logans.actuator.abs.AbstractTaskActuator;
+import com.bkcc.logans.constant.LogansMQConstant;
+import com.bkcc.logans.constant.TaskConstant;
 import com.bkcc.logans.controller.base.BaseController;
 import com.bkcc.logans.dispatch.abs.AbstractTaskDispatch;
 import com.bkcc.logans.entity.TaskEntity;
+import com.bkcc.logans.enums.TaskStatusEnum;
 import com.bkcc.logans.handler.TaskHandler;
 import com.bkcc.logans.service.TaskService;
 import com.github.pagehelper.PageInfo;
@@ -14,8 +18,10 @@ import io.swagger.annotations.ApiImplicitParams;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
+import org.apache.activemq.command.ActiveMQQueue;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.jms.core.JmsMessagingTemplate;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -25,63 +31,75 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.ModelAndView;
 
+import javax.jms.Destination;
+import java.util.HashMap;
+import java.util.Map;
+
 /**
  * 【描 述】：日志分析任务配置表Controller
  * 【环 境】：J2SE 1.8
  *
- *  @author         陈汝晗
- *  @version        v1.0 新建
- *  @since          2019-09-23 11:00:57
+ * @author 陈汝晗
+ * @version v1.0 新建
+ * @since 2019-09-23 11:00:57
  */
 @RestController
 @Api(value = "日志分析任务配置表Controller")
 @RequestMapping("/api/task")
-public class TaskController extends BaseController{
+public class TaskController extends BaseController {
 
     /**
      * 【描 述】：日志分析任务配置表业务接口
      *
-     *  @since  2019-09-23 11:00:57
+     * @since 2019-09-23 11:00:57
      */
     @Autowired
     private TaskService taskService;
 
     /**
+     * 【描 述】：消息通知
+     *
+     * @since Aug 16, 2019 v1.0
+     */
+    @Autowired
+    private JmsMessagingTemplate jmsTemplate;
+
+    /**
      * 【描 述】：查询日志分析任务配置表信息列表
      *
-     * @since  2019-09-23 11:00:57
+     * @since 2019-09-23 11:00:57
      */
     @ApiOperation(value = "查询日志分析任务配置表信息列表")
     @ApiImplicitParams({
-        @ApiImplicitParam(paramType="body", name="task", value="{"
-            + "</br> pageNum : 当前页码--不传查询全部,"
-            + "</br> pageSize : 每页大小--不传查询全部"
-            + "</br> taskName : 分析任务中文名称,"
-            + "</br> moduleName : 微服务模块名称,"
-            + "</br> reqMethod : 需要分析请求的方法 GET POST DELETE PUT,"
-            + "</br> reqUri : 需要分析请求的URI,"
-            + "</br> ansType : 分析类型，1--不分析直接输出指定字段，2--通过指定字段聚合分析,"
-            + "</br> ansRateType : 分析频次，每（1--分，2--时，3--日，4--月）分析一次。,"
-            + "</br> outQueue : 分析结果输出消息队列名称,"
-            + "</br>}", dataType="TaskEntity", required=true)
+            @ApiImplicitParam(paramType = "body", name = "task", value = "{"
+                    + "</br> pageNum : 当前页码--不传查询全部,"
+                    + "</br> pageSize : 每页大小--不传查询全部"
+                    + "</br> taskName : 分析任务中文名称,"
+                    + "</br> moduleName : 微服务模块名称,"
+                    + "</br> reqMethod : 需要分析请求的方法 GET POST DELETE PUT,"
+                    + "</br> reqUri : 需要分析请求的URI,"
+                    + "</br> ansType : 分析类型，1--不分析直接输出指定字段，2--通过指定字段聚合分析,"
+                    + "</br> ansRateType : 分析频次，每（1--分，2--时，3--日，4--月）分析一次。,"
+                    + "</br> outQueue : 分析结果输出消息队列名称,"
+                    + "</br>}", dataType = "TaskEntity", required = true)
     })
     @ApiResponses({
-        @ApiResponse(code = 200, message = "{</br> rows:[{"
-            + "</br> id : 日志分析任务配置表ID,"
-            + "</br> taskName : 分析任务中文名称,"
-            + "</br> moduleName : 微服务模块名称,"
-            + "</br> reqMethod : 需要分析请求的方法 GET POST DELETE PUT,"
-            + "</br> reqUri : 需要分析请求的URI,"
-            + "</br> ansType : 分析类型，1--不分析直接输出指定字段，2--通过指定字段聚合分析,"
-            + "</br> ansRateType : 分析频次，每（1--分，2--时，3--日，4--月）分析一次。,"
-            + "</br> outQueue : 分析结果输出消息队列名称,"
-            + "</br>}...], "
-            + "</br> newPrimaryKeys : {},"
-            + "</br> total:总数 </br>}")
+            @ApiResponse(code = 200, message = "{</br> rows:[{"
+                    + "</br> id : 日志分析任务配置表ID,"
+                    + "</br> taskName : 分析任务中文名称,"
+                    + "</br> moduleName : 微服务模块名称,"
+                    + "</br> reqMethod : 需要分析请求的方法 GET POST DELETE PUT,"
+                    + "</br> reqUri : 需要分析请求的URI,"
+                    + "</br> ansType : 分析类型，1--不分析直接输出指定字段，2--通过指定字段聚合分析,"
+                    + "</br> ansRateType : 分析频次，每（1--分，2--时，3--日，4--月）分析一次。,"
+                    + "</br> outQueue : 分析结果输出消息队列名称,"
+                    + "</br>}...], "
+                    + "</br> newPrimaryKeys : {},"
+                    + "</br> total:总数 </br>}")
     })
     @PostMapping("/list")
     public ViewData selectTaskList(@RequestBody TaskEntity task) {
-        if(task == null) {
+        if (task == null) {
             return ViewData.error("传入信息有误");
         }
         PageInfo<TaskEntity> pageInfo = taskService.selectTaskList(task);
@@ -90,39 +108,61 @@ public class TaskController extends BaseController{
         returnV.settotal(pageInfo.getTotal());
         return returnV;
     }
-            
+
     /**
      * 【描 述】：添加、修改日志分析任务配置表信息
      *
-     * @since  2019-09-23 11:00:57
+     * @since 2019-09-23 11:00:57
      */
     @ApiOperation(value = "添加、修改日志分析任务配置表信息")
     @ApiImplicitParams({
-        @ApiImplicitParam(paramType="body", name="task", value="{"
-                + "</br> \"id\" : 主键ID，不传或-1代表新增,"
-                + "</br> \"taskName\" : 分析任务中文名称,"
-                + "</br> \"moduleName\" : 微服务模块名称,"
-                + "</br> \"reqMethod\" : 需要分析请求的方法 GET POST DELETE PUT,"
-                + "</br> \"reqUri\" : 需要分析请求的URI,"
-                + "</br> \"ansType\" : 分析类型，1--不分析直接输出指定字段，2--通过指定字段聚合分析,"
-                + "</br> \"ansRateType\" : 分析频次，每（1--分，2--时，3--日，4--月）分析一次。,"
-                + "</br> \"outQueue\" : 分析结果输出消息队列名称,"
-                + "</br>}", dataType="TaskEntity", required=true)
+            @ApiImplicitParam(paramType = "body", name = "task", value = "{"
+                    + "</br> \"id\" : 主键ID，不传或-1代表新增,"
+                    + "</br> \"taskName\" : 分析任务中文名称,"
+                    + "</br> \"moduleName\" : 微服务模块名称,"
+                    + "</br> \"reqMethod\" : 需要分析请求的方法 GET POST DELETE PUT,"
+                    + "</br> \"reqUri\" : 需要分析请求的URI,"
+                    + "</br> \"ansType\" : 分析类型，1--不分析直接输出指定字段，2--通过指定字段聚合分析,"
+                    + "</br> \"ansRateType\" : 分析频次，每（1--分，2--时，3--日，4--月）分析一次。,"
+                    + "</br> \"outQueue\" : 分析结果输出消息队列名称,"
+                    + "</br>}", dataType = "TaskEntity", required = true)
     })
     @ApiResponses({
-        @ApiResponse(code = 200, message = "{</br>newPrimaryKeys : {"
-            + "</br> id : 日志分析任务配置表ID"
-            + "</br>}</br>}" )
+            @ApiResponse(code = 200, message = "{</br>newPrimaryKeys : {"
+                    + "</br> id : 日志分析任务配置表ID"
+                    + "</br>}</br>}")
     })
     @PostMapping
     public ViewData insertOrUpdate(@RequestBody TaskEntity task) {
-        if(task == null) {
+        if (task == null) {
             return ViewData.error("传入信息有误");
         }
+        if (task.getStatus() == null) {
+            task.setStatus(TaskStatusEnum.OPEN.getValue());
+        }
+        boolean isAdd = false;
+        if (task.getId() == null || task.getId() <= 0) {
+            isAdd = true;
+        }
         taskService.insertOrUpdate(task);
+
+        Map<String, TaskEntity> taskMap = new HashMap<>();
+        if (isAdd) {
+            if (TaskStatusEnum.OPEN.equels(task.getStatus())) {
+                taskMap.put(TaskConstant.INSERT_STATUS, task);
+            }
+        } else {
+            if (TaskStatusEnum.CLOSE.equels(task.getStatus())) {
+                taskMap.put(TaskConstant.DELETE_STATUS, task);
+            } else if (TaskStatusEnum.OPEN.equels(task.getStatus())) {
+                taskMap.put(TaskConstant.UPDATE_STATUS, task);
+            }
+        }
+        Destination destination = new ActiveMQQueue(LogansMQConstant.LOGANS_TASK);
+        jmsTemplate.convertAndSend(destination, JSONObject.toJSONString(taskMap));
         return ViewData.put("id", task.getId());
     }
-    
+
     /**
      * 【描 述】：删除日志分析任务配置表信息
      *
@@ -132,14 +172,22 @@ public class TaskController extends BaseController{
      */
     @ApiOperation(value = "删除日志分析任务配置表信息")
     @ApiImplicitParams({
-        @ApiImplicitParam(paramType="path", name="taskId", value="日志分析任务配置表ID", dataType="Long", required=true)
+            @ApiImplicitParam(paramType = "path", name = "taskId", value = "日志分析任务配置表ID", dataType = "Long", required = true)
     })
     @DeleteMapping("/{taskId}/delete")
     public ViewData deleteById(@PathVariable Long taskId) {
+        TaskEntity task = taskService.selectTaskById(taskId);
+        if (task == null) {
+            return ViewData.ok();
+        }
         taskService.deleteTaskById(taskId);
+        Map<String, TaskEntity> taskMap = new HashMap<>();
+        taskMap.put(TaskConstant.DELETE_STATUS, task);
+        Destination destination = new ActiveMQQueue(LogansMQConstant.LOGANS_TASK);
+        jmsTemplate.convertAndSend(destination, JSONObject.toJSONString(taskMap));
         return ViewData.ok();
     }
-    
+
     /**
      * 【描 述】：查询单个日志分析任务配置表信息
      *
@@ -149,19 +197,19 @@ public class TaskController extends BaseController{
      */
     @ApiOperation(value = "查询单个日志分析任务配置表信息")
     @ApiImplicitParams({
-        @ApiImplicitParam(paramType="path", name="taskId", value="日志分析任务配置表ID", dataType="Long", required=true)
+            @ApiImplicitParam(paramType = "path", name = "taskId", value = "日志分析任务配置表ID", dataType = "Long", required = true)
     })
     @ApiResponses({
-        @ApiResponse(code = 200, message = "{</br> rows:[{"
-                + "</br> taskName : 分析任务中文名称,"
-                + "</br> moduleName : 微服务模块名称,"
-                + "</br> reqMethod : 需要分析请求的方法 GET POST DELETE PUT,"
-                + "</br> reqUri : 需要分析请求的URI,"
-                + "</br> ansType : 分析类型，1--不分析直接输出指定字段，2--通过指定字段聚合分析,"
-                + "</br> ansRateType : 分析频次，每（1--分，2--时，3--日，4--月）分析一次。,"
-                + "</br> outQueue : 分析结果输出消息队列名称,"
-            + "</br>}], </br> newPrimaryKeys : {}"
-            + "</br>}" )
+            @ApiResponse(code = 200, message = "{</br> rows:[{"
+                    + "</br> taskName : 分析任务中文名称,"
+                    + "</br> moduleName : 微服务模块名称,"
+                    + "</br> reqMethod : 需要分析请求的方法 GET POST DELETE PUT,"
+                    + "</br> reqUri : 需要分析请求的URI,"
+                    + "</br> ansType : 分析类型，1--不分析直接输出指定字段，2--通过指定字段聚合分析,"
+                    + "</br> ansRateType : 分析频次，每（1--分，2--时，3--日，4--月）分析一次。,"
+                    + "</br> outQueue : 分析结果输出消息队列名称,"
+                    + "</br>}], </br> newPrimaryKeys : {}"
+                    + "</br>}")
     })
     @GetMapping("/{taskId}/detail")
     public ViewData selectTaskById(@PathVariable Long taskId) {
@@ -202,7 +250,6 @@ public class TaskController extends BaseController{
     }
 
 
-
     @Autowired
     @Qualifier("pollTaskDispatch")
     private AbstractTaskDispatch taskDispatch;
@@ -212,7 +259,6 @@ public class TaskController extends BaseController{
     private AbstractTaskActuator actuator;
 
 
-
     @GetMapping("/{taskId}/trigger")
     public ViewData trigger(@PathVariable Long taskId) {
         actuator.setTaskId(taskId);
@@ -220,7 +266,6 @@ public class TaskController extends BaseController{
         new Thread(taskHandler).start();
         return ViewData.ok();
     }
-
 
 
 }///：～
